reorganize ui files
