+++ lib/app/app.dart
@@ -2,14 +2,13 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chatbot/app/models/theme_model.dart';
 import 'package:flutter_chatbot/app/services/firebase_auth_service.dart';
-import 'package:flutter_chatbot/ui/views/authentication/sign_in/create_account_view.dart';
-// import 'package:flutter_chatbot/ui/views/home/home_view.dart';
+import 'package:flutter_chatbot/ui/views/authentication/create_account_view.dart';
 import 'package:flutter_chatbot/ui/views/onboarding/onboard_pageview.dart';
-import 'package:flutter_chatbot/ui/views/private_route.dart';
-import 'package:flutter_chatbot/ui/views/public_route.dart';
+import 'package:flutter_chatbot/app/routing/private_route.dart';
+import 'package:flutter_chatbot/app/routing/public_route.dart';
 import 'package:provider/provider.dart';
 import '../app/constants/strings.dart';
-import '../ui/views/authentication/sign_in/sign_in_view.dart';
+import '../ui/views/authentication/sign_in_view.dart';
 import '../ui/views/messaging/messaging_view.dart';
 
 class MyApp extends StatelessWidget {
@@ -30,7 +29,6 @@ class MyApp extends StatelessWidget {
               Strings.createAccountRoute: (_) => PublicRoute(
                     route: CreateAccountView(),
                   ),
-              // Strings.homeRoute: (_) => PrivateRoute(route: HomeView()),
               Strings.messagingViewRoute: (_) =>
                   PrivateRoute(route: MessagingView()),
               Strings.onBoardingRoute: (_) =>
