+++ lib/ui/views/auth/sign_in_view.dart
@@ -1,11 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chatbot/app/constants/strings.dart';
-import 'package:flutter_chatbot/ui/views/authentication/sign_in/widgets/auth_view_header.dart';
-import 'package:flutter_chatbot/ui/views/authentication/sign_in/widgets/create_account_form.dart';
-
+import 'package:flutter_chatbot/ui/widgets/auth/core/auth_view_header.dart';
+import 'package:flutter_chatbot/ui/widgets/auth/forms/sign_in_form.dart';
 import 'package:flutter_svg/svg.dart';
+import '../../widgets/auth/core/google_sign_in_button.dart';
 
-class CreateAccountView extends StatelessWidget {
+class SignInView extends StatelessWidget {
+  const SignInView({Key key}) : super(key: key);
+
+  @override
   Widget build(BuildContext context) {
     return Container(
       color: Theme.of(context).backgroundColor,
@@ -22,16 +25,22 @@ class CreateAccountView extends StatelessWidget {
                 mainAxisAlignment: MainAxisAlignment.start,
                 children: [
                   AuthViewHeader(),
-                  CreateAccountForm(),
+                  SignInForm(),
+                  Container(
+                    margin: EdgeInsets.only(top: 20),
+                    child: SvgPicture.asset("assets/images/OR.svg"),
+                  ),
+                  GoogleSignInButton(),
                   Spacer(),
                   Container(
                       margin: EdgeInsets.only(bottom: 20),
                       child: FlatButton(
                         onPressed: () {
-                          Navigator.pop(context);
+                          Navigator.pushNamed(
+                              context, Strings.createAccountRoute);
                         },
                         child: Text(
-                          "Already have an account? Login",
+                          "New user? Create an account",
                           style: Theme.of(context)
                               .textTheme
                               .bodyText1
