--- lib/ui/views/authentication/sign_in/widgets/anonymous_sign_in_button.dart
@@ -1,30 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_chatbot/app/services/firebase_auth_service.dart';
-import 'package:provider/provider.dart';
-
-import '../../../../../app/constants/strings.dart';
-
-class AnonymousSignInButton extends StatelessWidget {
-  const AnonymousSignInButton({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return RaisedButton(
-      onPressed: () {
-        Provider.of<AuthService>(context, listen: false).signInAnonymously();
-      },
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-      color: Colors.red[900],
-      child: Text(
-        Strings.anonymousSignUp,
-        textAlign: TextAlign.center,
-        style: TextStyle(
-          fontSize: 14,
-          color: Colors.white,
-        ),
-      ),
-    );
-  }
-}
