+++ lib/ui/widgets/messaging/chatbot/avatar_view.dart
