+++ lib/ui/widgets/messaging/chatbot/bot_response.dart
