+++ lib/ui/widgets/messaging/core/text_composer.dart
