+++ lib/ui/widgets/messaging/message/chat_message.dart
@@ -129,64 +129,3 @@ class ChatMessage extends StatelessWidget {
     );
   }
 }
-
-// child: Consumer<ChatModel>(builder: (context, chat, child) {
-//                   return Row(
-//                     children: [
-//                       GestureDetector(
-//                         onTap: () {
-//                           Provider.of<ChatModel>(context, listen: false)
-//                               .changeSelected(index);
-//                         },
-//                         child: Bubble(
-//                           child: Text(text),
-//                           color: Color.fromRGBO(225, 225, 225, 1.0),
-//                         ),
-//                       ),
-//                       if ((index == chat.getChatList().length - 1) | selected)
-//                         IconButton(
-//                           icon: FaIcon(
-//                             FontAwesomeIcons.check,
-//                             color: chat.getChatList()[index].feedback == 0
-//                                 ? Colors.green
-//                                 : Colors.grey,
-//                           ),
-//                           iconSize: 15,
-//                           onPressed: () {
-//                             Provider.of<ChatModel>(context, listen: false)
-//                                 .giveFeedback(index, 0);
-//                           },
-//                         ),
-//                       if ((index == chat.getChatList().length - 1) | selected)
-//                         IconButton(
-//                           icon: FaIcon(
-//                             FontAwesomeIcons.times,
-//                             color: chat.getChatList()[index].feedback == 1
-//                                 ? Colors.red
-//                                 : Colors.grey,
-//                           ),
-//                           iconSize: 15,
-//                           onPressed: () {
-//                             Provider.of<ChatModel>(context, listen: false)
-//                                 .giveFeedback(index, 1);
-//                           },
-//                         ),
-//                       if (((index == chat.getChatList().length - 1) |
-//                               selected) &
-//                           (chat.getChatList()[index].feedback != -1))
-//                         IconButton(
-//                           icon: FaIcon(
-//                             FontAwesomeIcons.comment,
-//                             color: chat.getChatList()[index].comment.isEmpty
-//                                 ? Colors.grey
-//                                 : Colors.lightBlue,
-//                           ),
-//                           iconSize: 15,
-//                           onPressed: () {
-//                             Provider.of<ChatModel>(context, listen: false)
-//                                 .addComment(index);
-//                           },
-//                         )
-//                     ],
-//                   );
-//                 })),
