+++ lib/ui/widgets/messaging/message/chat_nip.dart
